🐛  - 라우팅 설정 및 파일 위치 정리
