--- frontend/src/pages/LoginPage/MyPage.tsx
@@ -1,15 +1,0 @@
-import { css } from '@emotion/react';
-import MyPageTemplate from '../../components/templates/Login/MyPageTemplate';
-
-export default function MyPage() {
-    return (
-        <div css={wrapperStyles}>
-            <MyPageTemplate />
-        </div>
-    );
-}
-
-const wrapperStyles = css`
-    width: 75rem;
-    margin: auto;
-`;
